+++ TaskSchedulingCPP/slave_computing.hpp
@@ -1,7 +1,7 @@
 #ifndef SLAVE_COMPUTING_HEADER
 #define SLAVE_COMPUTING_HEADER
 
-bool GetCommunicationParameter(const double *, size_t, CEESParameter *); 
+bool GetCommunicationParameter(const double *, int, CEESParameter *); 
 // CEESParameter cannot be const, because its h and t will be altered upon the received message
 
 std::vector<int> ExecutingSimulationTask(TDenseMatrix &jump_table, CEquiEnergyModel &model, int my_rank, int group_index, int nGroup, const CSampleIDWeight &mode, int message_tag);
